% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datas.R
\docType{data}
\name{xy}
\alias{xy}
\title{My data}
\format{
A \code{data.frame} with 100 rows and 2 columns:
\itemize{
\item \code{x} is a random sample of normal distribution with R function "rnorm"
\item \code{y} is a random sample of normal distribution with R function "rnorm"
}
}
\source{
\url{https://en.wikipedia.org/wiki/Data}
}
\usage{
xy
}
\description{
This is my first data !
}
\references{
\url{https://en.wikipedia.org/wiki/Data}
}
\author{
Arnaud Lemercinier \href{mailto:arnaud.lemercinier@gmail.com}{arnaud.lemercinier@gmail.com}
}
\keyword{datasets}
